+++ colossalai/context/__init__.py
@@ -1,6 +1,5 @@
 from .config import Config, ConfigException
 from .parallel_context import ParallelContext
 from .parallel_mode import ParallelMode
-from .moe_context import MOE_CONTEXT
 from .process_group_initializer import *
 from .random import *
