+++ colossalai/nn/layer/__init__.py
@@ -4,7 +4,6 @@ from .parallel_2d import *
 from .parallel_2p5d import *
 from .parallel_3d import *
 from .parallel_sequence import *
-from .moe import *
 from .utils import *
 from .vanilla import *
 from .wrapper import *
